# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: phpGRC
  version: 0.4.7
  description: |
    Notes:
    - The Metrics Dashboard is an internal, admin-only endpoint (`GET /dashboard/kpis`, alias `GET /metrics/dashboard`) and is not part of this public OpenAPI. Both return identical shapes.
  license:
    identifier: MIT
    name: MIT
  x-logo:
    url: "https://phpgrc.gruntlabs.net/api/images/phpGRC-light-horizontal-trans.png"
    altText: "phpGRC"
    href: "https://phpgrc.gruntlabs.net/api/docs"
servers:
  - url: /
...
security:
  - sanctum: []
paths:
...
  /audit/categories:
    get:
      operationId: auditCategories
      tags: [Audit]
      summary: List audit categories
      description: Returns the fixed set of audit categories the system uses.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok, categories]
                properties:
                  ok:
                    type: boolean
                    const: true
                  categories:
                    type: array
                    items:
                      type: string
                      enum: [SYSTEM, RBAC, AUTH, SETTINGS, EXPORTS, EVIDENCE, AVATARS, AUDIT]
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: Token
...
  responses:
    Unauthenticated:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Unauthenticated'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ForbiddenError'
...
    RateLimited:
      description: Rate limited
      headers:
        Retry-After:
          description: Seconds until a new request is allowed.
          schema:
            type: integer
            minimum: 1
        X-RateLimit-Limit:
          description: Requests allowed in the window.
          schema:
            type: integer
            minimum: 1
        X-RateLimit-Remaining:
          description: Requests remaining in the current window.
          schema:
            type: integer
            minimum: 0
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimited'
  schemas:
...
    ErrorSimple:
      type: object
      required: [ok, code]
      properties:
        ok:
          type: boolean
          const: false
        code:
          type: string
        message:
          oneOf:
            - type: string
            - type: 'null'
        note:
          oneOf:
            - type: string
            - type: 'null'
    Unauthenticated:
      allOf:
        - $ref: '#/components/schemas/ErrorSimple'
        - type: object
          properties:
            code:
              type: string
              enum: [UNAUTHENTICATED]
    ForbiddenError:
      allOf:
        - $ref: '#/components/schemas/ErrorSimple'
        - type: object
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED]
...
    RateLimited:
      type: object
      required: [ok, code, retry_after]
      properties:
        ok:
          type: boolean
          const: false
        code:
          type: string
          enum: [RATE_LIMITED]
        retry_after:
          type: integer
          minimum: 1
...